Extract NavBar layout, solidify user dropdown, add side-nav sign out
