+++ src/components/pages/DatasourcesPage.css
@@ -2,17 +2,7 @@
 
 /* Wrapper tweaks */
 .data-sources-page {
-  display: grid;
-  grid-template-columns: 1fr;
-  gap: 20px;
-  align-items: flex-start;
-}
-
-.dashboard-main {
-  display: grid;
-  grid-template-columns: 1fr;
-  gap: 20px;
-  align-items: flex-start;
+  padding-bottom: 40px;
 }
 
 /* HERO – like a big card */
